requirements.txt
flask==3.0.0
tavily-python==0.3.0
python-dotenv==1.0.0
tiktoken==0.5.1
requests==2.31.0
